functions/oldrequirements.txt
functions-framework
google-cloud-firestore
firebase-functions
oauth2client
pandas
matplotlib
seaborn
